Move 'dialogs' package into 'view' package
